Add validation-error.js and validation-rules.js. Add exports for functions and class and imports.
