+++ fileHub/js-app/app/validation-error.js
@@ -1,0 +1,7 @@
+export default class ValidationError {
+
+    constructor(idElement, massage) {
+        this.idElement = idElement;
+        this.massage = massage;
+    }
+}
